allocatePinnedHostMemory(): use size_t.
This allows larger than 2GB buffers to be allocated for the GPU.
Also use size_t in GPU Transfer functions.

+++ HostCUDA.cu
@@ -62,7 +62,7 @@ extern "C" double CmiWallTimer();
 #endif
 
 
-void allocatePinnedHostMemory(void **ptr, int size){
+void allocatePinnedHostMemory(void **ptr, size_t size){
   if(size <= 0){
     *((char **)ptr) = NULL;
 #ifdef CUDA_PRINT_ERRORS
@@ -77,7 +77,7 @@ void allocatePinnedHostMemory(void **ptr, int size){
   cudaMallocHost(ptr, size);
 #endif
 #ifdef CUDA_PRINT_ERRORS
-  printf("allocatePinnedHostMemory: %s size: %d\n", cudaGetErrorString( cudaGetLastError() ), size);
+  printf("allocatePinnedHostMemory: %s size: %zu\n", cudaGetErrorString( cudaGetLastError() ), size);
 #endif
 }
 
@@ -275,7 +275,7 @@ void DataManagerTransferLocalTree(CudaMultipoleMoments *moments, int nMoments,
 #endif
 	transferKernel->setRunKernel(run_DM_TRANSFER_LOCAL);
 #ifdef CUDA_VERBOSE_KERNEL_ENQUEUE
-        printf("(%d) DM LOCAL TREE moments %d (%d) partcores %d (%d) partvars %d (%d)\n",
+        printf("(%d) DM LOCAL TREE moments %zu (%d) partcores %zu (%d) partvars %zu (%d)\n",
                   CmiMyPe(),
                   transferKernel->buffers[LOCAL_MOMENTS_IDX].size,
                   transferKernel->buffers[LOCAL_MOMENTS_IDX].transfer_to_device,
@@ -335,7 +335,7 @@ void DataManagerTransferLocalTree(CudaMultipoleMoments *moments, int nMoments,
                                 REMOTE_PARTICLE_CORES);
 
 #ifdef CUDA_VERBOSE_KERNEL_ENQUEUE
-  printf("(%d) DM REMOTE CHUNK moments %d (%d) partcores %d (%d)\n",
+  printf("(%d) DM REMOTE CHUNK moments %zu (%d) partcores %zu (%d)\n",
             CmiMyPe(),
             transferKernel->buffers[REMOTE_MOMENTS_IDX].size,
             transferKernel->buffers[REMOTE_MOMENTS_IDX].transfer_to_device,
@@ -786,7 +786,7 @@ void TreePieceCellListDataTransferRemote(CudaRequest *data){
 
 void TreePieceCellListDataTransferRemoteResume(CudaRequest *data, CudaMultipoleMoments *missedMoments, int numMissedMoments){
   int numBlocks = data->numBucketsPlusOne-1;
-  int size;
+  size_t size;
 
   hapiWorkRequest* gravityKernel = hapiCreateWorkRequest();
   void* bufferHostBuffer;
@@ -835,7 +835,7 @@ void TreePieceCellListDataTransferRemoteResume(CudaRequest *data, CudaMultipoleM
 void TreePieceCellListDataTransferBasic(CudaRequest *data, hapiWorkRequest *gravityKernel){
 	int numBucketsPlusOne = data->numBucketsPlusOne;
         int numBuckets = numBucketsPlusOne-1;
-        int size = (data->numInteractions) * sizeof(ILCell);
+        size_t size = (data->numInteractions) * sizeof(ILCell);
         bool transfer = size > 0;
 
 	gravityKernel->addBuffer(data->list, size, transfer, false, transfer);
@@ -868,7 +868,7 @@ void TreePieceCellListDataTransferBasic(CudaRequest *data, hapiWorkRequest *grav
         gravityKernel->setUserData(ptr, true);
 
 #ifdef CUDA_VERBOSE_KERNEL_ENQUEUE
-        printf("(%d) TRANSFER BASIC cells %d (%d) bucket_markers %d (%d) bucket_starts %d (%d) bucket_sizes %d (%d)\n",
+        printf("(%d) TRANSFER BASIC cells %zu (%d) bucket_markers %zu (%d) bucket_starts %zu (%d) bucket_sizes %zu (%d)\n",
             CmiMyPe(),
             gravityKernel->buffers[ILCELL_IDX].size,
             gravityKernel->buffers[ILCELL_IDX].transfer_to_device,
@@ -903,7 +903,7 @@ void TreePiecePartListDataTransferLocalSmallPhase(CudaRequest *data, CompactPart
         size = (len) * sizeof(CompactPartData);
 
 #ifdef CUDA_VERBOSE_KERNEL_ENQUEUE
-        printf("(%d) TRANSFER LOCAL SMALL PHASE  %d (%d)\n",
+        printf("(%d) TRANSFER LOCAL SMALL PHASE  %zu (%d)\n",
             CmiMyPe(),
             size,
             transfer
@@ -1013,7 +1013,7 @@ void TreePiecePartListDataTransferRemoteResume(CudaRequest *data, CompactPartDat
         size = (numMissedParts) * sizeof(CompactPartData);
 
 #ifdef CUDA_VERBOSE_KERNEL_ENQUEUE
-        printf("(%d) TRANSFER REMOTE RESUME PART %d (%d)\n",
+        printf("(%d) TRANSFER REMOTE RESUME PART %zu (%d)\n",
             CmiMyPe(),
             size,
             transfer
@@ -1048,7 +1048,7 @@ void TreePiecePartListDataTransferBasic(CudaRequest *data, hapiWorkRequest *grav
 	int numInteractions = data->numInteractions;
 	int numBucketsPlusOne = data->numBucketsPlusOne;
         int numBuckets = numBucketsPlusOne-1;
-        int size;
+        size_t size;
         bool transfer;
 
 	size = (numInteractions) * sizeof(ILCell);
@@ -1071,7 +1071,7 @@ void TreePiecePartListDataTransferBasic(CudaRequest *data, hapiWorkRequest *grav
         gravityKernel->setUserData(ptr, true);
 
 #ifdef CUDA_VERBOSE_KERNEL_ENQUEUE
-        printf("(%d) TRANSFER BASIC PART parts %d (%d) bucket_markers %d (%d) bucket_starts %d (%d) bucket_sizes %d (%d)\n",
+        printf("(%d) TRANSFER BASIC PART parts %zu (%d) bucket_markers %zu (%d) bucket_starts %zu (%d) bucket_sizes %zu (%d)\n",
             CmiMyPe(),
             gravityKernel->buffers[ILPART_IDX].size,
             gravityKernel->buffers[ILPART_IDX].transfer_to_device,
@@ -1183,7 +1183,7 @@ void TransferParticleVarsBack(VariablePartData *hostBuffer, int size, void *cb,
      bool freemom, bool freepart, bool freeRemoteMom, bool, freeRemotePart,
      int index, char phase){
 #else
-void TransferParticleVarsBack(VariablePartData *hostBuffer, int size, void *cb,
+void TransferParticleVarsBack(VariablePartData *hostBuffer, size_t size, void *cb,
      bool freemom, bool freepart, bool freeRemoteMom, bool freeRemotePart){
 #endif
   hapiWorkRequest* gravityKernel = hapiCreateWorkRequest();
